LayoutTests/platform/chromium-mac/fast/canvas/canvas-draw-canvas-on-canvas-shadow-expected.txt

Ensure correct behavior when drawing a canvas on a canvas with shadows. A blue and red checkered pattern should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 127
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 127
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 127
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 100 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 100 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 100 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 50 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 50 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 50 (actual: 0)
PASS successfullyParsed is true

TEST COMPLETE
